requirements.txt
requests>=2.31.0,<3.0.0
PyJWT[crypto]>=2.10.1,<3.0.0
